wip traces interface
