+++ packages/core/src/storage/domains/index.ts
@@ -1,3 +1,4 @@
 export * from './scores';
 export * from './operations';
 export * from './workflows';
+export * from './traces';
