+++ packages/core/src/storage/domains/traces/index.ts
@@ -1,0 +1,2 @@
+export * from './base';
+export * from './inmemory';
